Prevent test_cnn.py from crashing w/invalid Tensor label
When an object is detected and then leaves the image, the Tensor label
returned is typically far outside the list of label indexes (ex: 7552.0). This
causes the `test_cnn.py` to crash.

This checks to ensure the index is within the available labels before
applying the label text.

The exception:

```
Traceback (most recent call last):
  File "test_cnn.py", line 85, in <module>
    cv2.putText(frame, labels[int(e[0]['label'])], pt_t1, cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 0, 255), 2)
IndexError: list index out of range
```