Allow use of prior cached model
Moving the requests.get(url) below the check for cache.  If your vision system is offline, and you have a cached model, the check prior to checking cache throws an exception.  Now if you have a cached model it will use that first and only if you dont have a model will it switch to checking api.roboflow.com